refactor(core): relocate AccessRequirement + group/system-env helpers to rise-backend-core
WS3 Phase 1 (PR 1B, increment 1): move the backend-agnostic pieces the Docker
controller depends on out of rise-deploy internals into rise-backend-core, behind
re-export shims so all existing crate:: paths keep working.

- AccessRequirement -> core::access (new `schema` feature gates the schemars
  JsonSchema derive, mirroring rise-deployment-spec; enabled by rise-deploy's
  backend feature). settings.rs re-exports it.
- DEFAULT_DEPLOYMENT_GROUP + normalize_deployment_group -> core::group.
- rise_system_env_vars -> core::system_env.
  Both re-exported from server::deployment::models.

Regenerated backend-settings.schema.json (one-line description text only).
Behavior-preserving: check + clippy --all-targets clean; 31 models tests pass.
